Flush each line of the response as it's written.
Chunks yielded from body.each may be generated on the fly, and we want
them to be written out as soon as possible, rather than waiting in
jetty's 8k response buffer.

This can help with html page load times as the header (likely less than
8k) can be sent ahead of the body being generated, meaning stylesheets
and js can be loaded by the client while the rest of the response is
still being produced.